[Minor] Rename: formating -> formatting (#1227)
